+++ SpaceTraders/Http/AgentService.cs
@@ -1,0 +1,17 @@
+using SpaceTraders.Models;
+using SpaceTraders.Services;
+namespace SpaceTraders.Http;
+
+public class AgentService(HttpClient httpClient) : BaseApiService(httpClient) {
+    public async Task<Agent?> GetAgentAsync() {
+        var deserializer = new Deserializer();
+        try {
+            using var jsonStream = await HttpClient.GetStreamAsync("https://api.spacetraders.io/v2/my/agent");
+            return await deserializer.DeserializeAgent(jsonStream);
+        }
+        catch (HttpRequestException ex) {
+            Console.WriteLine($"HTTP request to fetch agent data failed: {ex.Message}");
+            return null;
+        }
+    }
+}
