+++ SpaceTraders/Http/LocationService.cs
@@ -1,0 +1,33 @@
+using SpaceTraders.Models;
+using SpaceTraders.Services;
+namespace SpaceTraders.Http;
+
+public class LocationService(HttpClient httpClient) : BaseApiService(httpClient) {
+    public async Task<SystemDetails?> GetSystemAsync(string system) {
+        var deserializer = new Deserializer();
+        try {
+            await using var jsonStream =
+                await HttpClient.GetStreamAsync($"https://api.spacetraders.io/v2/systems/{system}");
+            return await deserializer.DeserializeSystem(jsonStream);
+        }
+        catch (HttpRequestException ex) {
+            Console.WriteLine($"HTTP request to fetch location data failed: {ex.Message}");
+            return null;
+        }
+    }
+
+    public async Task<Waypoint?> GetWaypointAsync(string waypoint) {
+        var deserializer = new Deserializer();
+        string system = waypoint.Substring(0, 7);
+        try {
+            await using var jsonStream =
+                await HttpClient.GetStreamAsync($"https://api.spacetraders.io/v2/systems/{system}/waypoints/{waypoint}");
+            return await deserializer.DeserializeWaypoint(jsonStream);
+        }
+        catch (HttpRequestException ex) {
+            Console.WriteLine($"HTTP request to fetch location data failed: {ex.Message}");
+            return null;
+        }
+    }
+    
+}
