+++ SpaceTraders/Models/Contract.cs
@@ -1,0 +1,50 @@
+using System.Text.Json.Serialization;
+
+namespace SpaceTraders.Models;
+
+public record Contracts( // rename ContractsList?
+    [property: JsonPropertyName("id")] string ContractID,
+    [property: JsonPropertyName("factionSymbol")]
+    string FactionSymbol,
+    [property: JsonPropertyName("type")] string ContractType,
+    [property: JsonPropertyName("terms")] Terms Terms,
+    [property: JsonPropertyName("accepted")]
+    bool Accepted
+);
+
+public record Contract( // todo define the rest of the properties
+    [property: JsonPropertyName("id")] string ContractID,
+    [property: JsonPropertyName("factionSymbol")]
+    string FactionSymbol,
+    [property: JsonPropertyName("type")] string ContractType,
+    [property: JsonPropertyName("terms")] Terms Terms,
+    [property: JsonPropertyName("accepted")]
+    bool Accepted
+);
+
+public record Terms(
+    [property: JsonPropertyName("deadline")]
+    DateTime Deadline,
+    [property: JsonPropertyName("payment")]
+    Payment Payment,
+    [property: JsonPropertyName("deliver")]
+    Deliver[] DeliverItems
+);
+
+public record Payment(
+    [property: JsonPropertyName("onAccepted")]
+    int OnAccepted,
+    [property: JsonPropertyName("onFulfilled")]
+    int OnFulfilled
+);
+
+public record Deliver(
+    [property: JsonPropertyName("tradeSymbol")]
+    string TradeSymbol,
+    [property: JsonPropertyName("destinationSymbol")]
+    string DestinationSymbol,
+    [property: JsonPropertyName("unitsRequired")]
+    int UnitsRequired,
+    [property: JsonPropertyName("unitsFulfilled")]
+    int UnitsFulfilled
+);
