--- SpaceTraders/Services/ContractService.cs
@@ -1,5 +1,0 @@
-namespace SpaceTraders;
-
-public class ContractService {
- 
-}
