+++ SpaceTraders/Services/Deserializer.cs
@@ -1,98 +1,13 @@
-﻿namespace SpaceTraders;
-
-using System;
+﻿using System;
 using System.IO;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using SpaceTraders.Models;
 
+namespace SpaceTraders.Services;
+
 public class Deserializer { // todo move all records to respective model files
-    public record Waypoint(
-        [property: JsonPropertyName("symbol")] string Symbol,
-        [property: JsonPropertyName("type")] string Type,
-        [property: JsonPropertyName("systemSymbol")]
-        string SystemSymbol,
-        [property: JsonPropertyName("x")] int X,
-        [property: JsonPropertyName("y")] int Y,
-        [property: JsonPropertyName("orbitals")]
-        Orbital[] Orbitals,
-        [property: JsonPropertyName("traits")] Trait[] Traits,
-        [property: JsonPropertyName("isUnderConstruction")]
-        bool IsUnderConstruction,
-        [property: JsonPropertyName("faction")]
-        Faction Faction,
-        [property: JsonPropertyName("modifiers")]
-        object[] Modifiers,
-        [property: JsonPropertyName("chart")] Chart Chart
-    );
-
-    public record Trait(
-        [property: JsonPropertyName("symbol")] string Symbol,
-        [property: JsonPropertyName("name")] string Name,
-        [property: JsonPropertyName("description")]
-        string Description
-    );
-
-    public record Faction(
-        [property: JsonPropertyName("symbol")] string Symbol
-    );
-
-    public record Chart(
-        [property: JsonPropertyName("waypointSymbol")]
-        string WaypointSymbol,
-        [property: JsonPropertyName("submittedBy")]
-        string SubmittedBy,
-        [property: JsonPropertyName("submittedOn")]
-        DateTime SubmittedOn
-    );
-
-    public record Contracts( // rename ContractsList?
-        [property: JsonPropertyName("id")] string ContractID,
-        [property: JsonPropertyName("factionSymbol")]
-        string FactionSymbol,
-        [property: JsonPropertyName("type")] string ContractType,
-        [property: JsonPropertyName("terms")] Terms Terms,
-        [property: JsonPropertyName("accepted")]
-        bool Accepted
-    );
-
-    public record Contract( // todo define the rest of the properties
-        [property: JsonPropertyName("id")] string ContractID,
-        [property: JsonPropertyName("factionSymbol")]
-        string FactionSymbol,
-        [property: JsonPropertyName("type")] string ContractType,
-        [property: JsonPropertyName("terms")] Terms Terms,
-        [property: JsonPropertyName("accepted")]
-        bool Accepted
-    );
-
-    public record Terms(
-        [property: JsonPropertyName("deadline")]
-        DateTime Deadline,
-        [property: JsonPropertyName("payment")]
-        Payment Payment,
-        [property: JsonPropertyName("deliver")]
-        Deliver[] DeliverItems
-    );
-
-    public record Payment(
-        [property: JsonPropertyName("onAccepted")]
-        int OnAccepted,
-        [property: JsonPropertyName("onFulfilled")]
-        int OnFulfilled
-    );
-
-    public record Deliver(
-        [property: JsonPropertyName("tradeSymbol")]
-        string TradeSymbol,
-        [property: JsonPropertyName("destinationSymbol")]
-        string DestinationSymbol,
-        [property: JsonPropertyName("unitsRequired")]
-        int UnitsRequired,
-        [property: JsonPropertyName("unitsFulfilled")]
-        int UnitsFulfilled
-    );
     
     public record Meta(
         [property: JsonPropertyName("total")] int Total,
